I *Think* I have a Dud Fix
So instead of removing all the duds (which takes too long), I store the number of matches in numMatches.

I grab a random artwork and if it's a dud AND numMatches . 1, I run the whole thing again to try and get a non-dud. We'll see if that works. So far, so good.

There's a chance that a particular query will have multiple results and ALL will be duds, which would create an infinite loop. So I need some sort of way to catch for that. Consulting with a pro later.

Next up, loading animation?